--- tests/integration/mnist.py
@@ -1,48 +1,0 @@
-"""
-MNIST Example
-"""
-
-#imports
-import os
-
-import torch as th
-import torch.nn as nn
-import torch.nn.functional as F
-from torchvision import datasets
-
-from torchsample.modules import ModuleTrainer
-from torchsample.callbacks import *
-from torchsample.regularizers import *
-from torchsample.constraints import *
-from torchsample.initializers import *
-from torchsample.metrics import *
-
-
-# LOAD DATA
-filepath = os.path.dirname(os.path.abspath(__file__))
-ROOT = os.path.join('/'.join(filepath.split('/')[:-1]),'/_data')
-dataset = datasets.MNIST(ROOT, train=True, download=True)
-
-x_train, y_train = th.load(os.path.join(dataset.root, 'processed/training.pt'))
-x_test, y_test = th.load(os.path.join(dataset.root, 'processed/test.pt'))
-
-x_train = x_train.float()
-y_train = y_train.long()
-x_test = x_test.float()
-y_test = y_test.long()
-
-x_train = x_train / 255.
-x_test = x_test / 255.
-x_train = x_train.unsqueeze(1)
-x_test = x_test.unsqueeze(1)
-
-# only train on a subset
-x_train = x_train[:10000]
-y_train = y_train[:10000]
-x_test = x_test[:1000]
-y_test = y_test[:1000]
-
-# ----------
-
-
-
