+++ tests/integration/simple_multi_input_no_target.py
@@ -23,8 +23,8 @@ x_train = x_train.unsqueeze(1)
 x_test = x_test.unsqueeze(1)
 
 # only train on a subset
-x_train = x_train[:10000]
-y_train = y_train[:10000]
+x_train = x_train[:1000]
+y_train = y_train[:1000]
 x_test = x_test[:1000]
 y_test = y_test[:1000]
 
@@ -36,44 +36,32 @@ class Network(nn.Module):
         self.conv1 = nn.Conv2d(1, 32, kernel_size=3)
         self.conv2 = nn.Conv2d(32, 64, kernel_size=3)
         self.fc1 = nn.Linear(1600, 128)
-        self.fc2 = nn.Linear(128, 10)
+        self.fc2 = nn.Linear(128, 1)
 
-    def forward(self, x):
+    def forward(self, x, y, z):
         x = F.relu(F.max_pool2d(self.conv1(x), 2))
         x = F.relu(F.max_pool2d(self.conv2(x), 2))
         x = x.view(-1, 1600)
         x = F.relu(self.fc1(x))
         x = F.dropout(x, training=self.training)
         x = self.fc2(x)
-        return F.log_softmax(x), F.log_softmax(x)
+        return th.abs(10 - x)
 
 
-# one loss function for multiple targets
 model = Network()
 trainer = ModuleTrainer(model)
-trainer.compile(loss='nll_loss',
+
+trainer.compile(loss='unconstrained_sum',
                 optimizer='adadelta')
 
-trainer.fit(x_train, 
-            [y_train, y_train], 
-            nb_epoch=1, 
+trainer.fit([x_train, x_train, x_train],
+            num_epoch=3, 
             batch_size=128,
             verbose=1)
-ypred1, ypred2 = trainer.predict(x_train)
-print(ypred1.size(), ypred2.size())
 
-eval_loss = trainer.evaluate(x_train, [y_train, y_train])
+ypred = trainer.predict([x_train, x_train, x_train])
+print(ypred.size())
+
+eval_loss = trainer.evaluate([x_train, x_train, x_train])
 print(eval_loss)
-# multiple loss functions
-model = Network()
-trainer = ModuleTrainer(model)
-trainer.compile(loss=['nll_loss', 'nll_loss'],
-                optimizer='adadelta')
-trainer.fit(x_train, 
-            [y_train, y_train], 
-            nb_epoch=3, 
-            batch_size=128,
-            verbose=1)
-
-
