+++ tests/integration/simple_single_input_multi_target.py
@@ -23,8 +23,8 @@ x_train = x_train.unsqueeze(1)
 x_test = x_test.unsqueeze(1)
 
 # only train on a subset
-x_train = x_train[:10000]
-y_train = y_train[:10000]
+x_train = x_train[:1000]
+y_train = y_train[:1000]
 x_test = x_test[:1000]
 y_test = y_test[:1000]
 
@@ -38,58 +38,42 @@ class Network(nn.Module):
         self.fc1 = nn.Linear(1600, 128)
         self.fc2 = nn.Linear(128, 10)
 
-    def forward(self, x, y, z):
+    def forward(self, x):
         x = F.relu(F.max_pool2d(self.conv1(x), 2))
         x = F.relu(F.max_pool2d(self.conv2(x), 2))
         x = x.view(-1, 1600)
         x = F.relu(self.fc1(x))
         x = F.dropout(x, training=self.training)
         x = self.fc2(x)
-        return F.log_softmax(x), F.log_softmax(x), F.log_softmax(x)
+        return F.log_softmax(x), F.log_softmax(x)
 
-# with one loss function given
+
+# one loss function for multiple targets
 model = Network()
 trainer = ModuleTrainer(model)
-
 trainer.compile(loss='nll_loss',
                 optimizer='adadelta')
 
-trainer.fit([x_train, x_train, x_train], 
-            [y_train, y_train, y_train],
-            nb_epoch=3, 
+trainer.fit(x_train, 
+            [y_train, y_train], 
+            num_epoch=3, 
             batch_size=128,
             verbose=1)
+ypred1, ypred2 = trainer.predict(x_train)
+print(ypred1.size(), ypred2.size())
 
-yp1, yp2, yp3 = trainer.predict([x_train, x_train, x_train])
-print(yp1.size(), yp2.size(), yp3.size())
-
-eval_loss = trainer.evaluate([x_train, x_train, x_train],
-                             [y_train, y_train, y_train])
+eval_loss = trainer.evaluate(x_train, [y_train, y_train])
 print(eval_loss)
-
-# With multiple loss functions given
+# multiple loss functions
 model = Network()
 trainer = ModuleTrainer(model)
-
-trainer.compile(loss=['nll_loss', 'nll_loss', 'nll_loss'],
-                optimizer='adadelta')
-
-trainer.fit([x_train, x_train, x_train], 
-            [y_train, y_train, y_train],
-            nb_epoch=3, 
-            batch_size=128,
-            verbose=1)
-
-# should raise exception for giving multiple loss functions 
-# but not giving a loss function for every input
-model = Network()
-trainer = ModuleTrainer(model)
-
 trainer.compile(loss=['nll_loss', 'nll_loss'],
                 optimizer='adadelta')
-
-trainer.fit([x_train, x_train, x_train], 
-            [y_train, y_train, y_train],
-            nb_epoch=3, 
+trainer.fit(x_train, 
+            [y_train, y_train], 
+            num_epoch=3, 
             batch_size=128,
             verbose=1)
+
+
+
